backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.16.1
docopt==0.6.2
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
google==3.0.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
openai==1.40.6
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==27.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
yarg==0.1.9
zipp==3.23.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
